pyproject.toml
[project]
name = "aetherpackbot"
version = "1.0.0"
description = "A modern, modular multi-platform chatbot framework with LLM support"
authors = [{ name = "AetherPackBot Team" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["chatbot", "llm", "telegram", "discord", "ai", "bot"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: AsyncIO",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "aiohttp>=3.9.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "orjson>=3.9.0",
    "tenacity>=8.2.0",
    "jinja2>=3.1.0",
    "aiosqlite>=0.19.0",
    "aiocache>=0.12.0",
    "croniter>=2.0.0",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
discord = ["discord.py>=2.3.0"]
openai = ["openai>=1.10.0"]
anthropic = ["anthropic>=0.18.0"]
google = ["google-generativeai>=0.4.0"]
all = [
    "python-telegram-bot>=21.0",
    "discord.py>=2.3.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
]
dev = [
    "ruff>=0.2.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
aetherpack = "aetherpackbot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
